requirements.txt
django
uagents
python-dotenv